- Move AutoComplete methods to trait - Move AutoCompleteConstants to Interfaces - Update twig errors
