+++ src/Fields/Select2HTMLMulti.php
@@ -6,6 +6,7 @@ use GCWorld\FormConfig\Core\Twig;
 use GCWorld\FormConfig\FieldInterface;
 use GCWorld\FormConfig\Forms\FormField;
 use GCWorld\FormConfig\Traits\Ajax;
+use GCWorld\FormConfig\Traits\AutoComplete;
 use GCWorld\FormConfig\Traits\Options;
 use GCWorld\FormConfig\Traits\Select2;
 
@@ -18,6 +19,7 @@ class Select2HTMLMulti extends Base implements FieldInterface
     use Ajax;
     use Options;
     use Select2;
+    use AutoComplete;
 
     /**
      * @return string
