+++ src/Forms/FormField.php
@@ -4,10 +4,10 @@ namespace GCWorld\FormConfig\Forms;
 use GCWorld\FormConfig\FieldInterface;
 use GCWorld\FormConfig\Generated\FieldConstants;
 use GCWorld\FormConfig\MultiSelectInterface;
+use GCWorld\FormConfig\Traits\AutoComplete;
 use GCWorld\FormConfig\Traits\FieldFormConfigTrait;
 use GCWorld\FormConfig\Traits\MetaDataTrait;
 use GCWorld\FormConfig\Traits\Select2;
-use GCWorld\FormConfig\Core\AutoCompleteConstants;
 
 /**
  * Class FormField.
@@ -17,6 +17,7 @@ class FormField
     use FieldFormConfigTrait;
     use MetaDataTrait;
     use Select2;
+    use AutoComplete;
 
     /**
      * @var array
@@ -144,11 +145,6 @@ class FormField
     protected $numberStep = null;
 
     /**
-     * @var null
-     */
-    protected $autoCompleteComponent = null;
-
-    /**
      * @param string $name
      */
     public function __construct(string $name)
@@ -233,26 +229,7 @@ class FormField
         return $this->name;
     }
 
-    public function getAutoComplete()
-    {
-        return $this->autoCompleteComponent;
-    }
 
-    /**
-     * @param $component
-     * @return $this
-     * @throws \Exception
-     */
-    public function setAutoComplete($component)
-    {
-        if (!isset(AutoCompleteConstants::COMPONENTS[$component])) {
-            $msg = 'Invalid Auto Complete Type: '.$component.'<br>Possible auto-complete types are: '.implode(', ',AutoCompleteConstants::COMPONENTS);
-            throw new \Exception($msg);
-        }
-
-        $this->autoCompleteComponent = $component;
-        return $this;
-    }
 
     /**
      * @param string $id
