+++ twig/BS3/field_parents/selectAjax.twig
@@ -6,7 +6,7 @@
             {{ multi ? 'multiple="multiple"':'' }}
             id="{{ field.id }}"
             class="form-control {{ field.class }}"
-            {{ (field.getAutoComplete() = '' ? 'autocomplete="off"' : 'autocomplete="' ~ field.getAutoComplete() ~ '"') }}
+            {{ (field.getAutoComplete() != '' ? 'autocomplete="' ~ field.getAutoComplete() ~ '"') }}
             {{ field.getDataAttributeString()|raw }}
             {{ field.labelledBy!= '' ? 'aria-labelledby="' ~ field.labelledby ~ '"' }}
     >
