+++ twig/BS3/fields/dateInput.twig
@@ -6,7 +6,7 @@
                value=""
                class="form-control {{ field.class }}"
                style="text-align:right"
-               {{ (field.getAutoComplete() = '' ? 'autocomplete="off"' : 'autocomplete="' ~ field.getAutoComplete()) ~ '"' }}
+               {{ (field.getAutoComplete()!= '' ? 'autocomplete="' ~ field.getAutoComplete() ~ '"' : 'autocomplete="off"')|raw }}
                placeholder="{{ field.getPlaceholder()|default('') }}"
                {{ (field.labelledBy!= '' ? 'aria-labelledby="' ~ field.labelledBy ~ '"')|raw }}
                 {{ field.getDataAttributeString()|raw }}
@@ -17,7 +17,7 @@
                name="{{ field.name }}"
                value="{{ field.value|date('m/d/Y') }}"
                style="text-align:right"
-               {{ (field.getAutoComplete() = '' ? 'autocomplete="off"' : 'autocomplete="' ~ field.getAutoComplete() ~ '"') }}
+               {{ (field.getAutoComplete()!= '' ? 'autocomplete="' ~ field.getAutoComplete() ~ '"' : 'autocomplete="off"')|raw }}
                class="form-control {{ field.class }}"
                placeholder="{{ field.getPlaceholder()|default('') }}"
                {{ (field.labelledBy!= '' ? 'aria-labelledby="' ~ field.labelledBy ~ '"')|raw }}
