+++ twig/BS3/fields/moneyInput.twig
@@ -10,6 +10,7 @@
            step="0.01"
            value="{{ field.value }}"
            onblur="$(this).val(parseFloat($(this).val()).toFixed(2))"
+           {{ (field.getAutoComplete() != '' ? 'autocomplete="' ~ field.getAutoComplete() ~ '"')|raw }}
            {{ (field.labelledBy!= '' ? 'aria-labelledby="' ~ field.labelledBy ~ '"')|raw }}
             {{ field.getDataAttributeString()|raw }}
     >
