+++ twig/BS3/fields/numberInput.twig
@@ -12,7 +12,7 @@
             maxlength="{{ field.maxLength }}"
             type="number"
             step="{{ step is null ? 'any' : step }}"
-            {{ (field.getAutoComplete() = '' ? 'autocomplete="off"' : 'autocomplete="' ~ field.getAutoComplete() ~ '"') }}
+            {{ (field.getAutoComplete() != '' ? 'autocomplete="' ~ field.getAutoComplete() ~ '"')|raw }}
             {% if min is not null  %} min="{{ min }}"{% endif %}
             {% if max is not null  %} max="{{ max }}"{% endif %}
             value="{{ field.value }}"
