+++ twig/BS3/fields/passwordInput.twig
@@ -5,7 +5,7 @@
                id="{{ field.id }}-input"
                value="" class="form-control {{ field.class }}"
                placeholder="{{ field.getPlaceholder()|default('') }}"
-               {{ (field.getAutoComplete() = '' ? 'autocomplete="off"' : 'autocomplete="' ~ field.getAutoComplete() ~ '"') }}
+               {{ (field.getAutoComplete() != '' ? 'autocomplete="' ~ field.getAutoComplete() ~ '"')|raw }}
                {{ (field.labelledBy!= '' ? 'aria-labelledby="' ~ field.labelledBy ~ '"')|raw }}
         >
         <span class="input-group-addon">Score: <span id="{{ field.id }}-score">0</span>/4</span>
