+++ twig/BS3/fields/textInput.twig
@@ -5,7 +5,7 @@
            value="{{ field.value|default('') }}"
            class="form-control {{ field.class }}"
            placeholder="{{ field.getPlaceholder()|default('') }}"
-           {{ (field.getAutoComplete() = '' ? 'autocomplete="off"' : 'autocomplete="' ~ field.getAutoComplete() ~ '"') }}
+           {{ (field.getAutoComplete() != '' ? 'autocomplete="' ~ field.getAutoComplete() ~ '"')|raw }}
            id="{{ field.id }}"
            {{ field.getDataAttributeString()|raw }}
            {{ (field.labelledBy!= '' ? 'aria-labelledby="' ~ field.labelledBy ~ '"')|raw }}
